Code review: comments, type safety, bounds checks, performance
CipPath: remove unused DataSegment constant.
CipInstance: XML docs on all members, cache sorted GetAll attribute list.
CipClass: XML docs, O(1) max instance tracking, AddInstance now updates max,
  stackalloc for attribute update instead of heap allocation.
CipDispatcher: XML docs, differentiated error codes per failure type
  (PathDestinationUnknown, ObjectDoesNotExist, ServiceNotSupported).
EncapsulationHeader: XML docs on all enum values and fields, Status field
  changed from uint to EncapsulationStatus enum, bounds check in Parse.
CpfItem: XML docs, bounds checks in Parse (truncated data handled gracefully),
  buffer size validation in Write.
MrCodec: XML docs, TryParseRequest with bool return, new TryParseResponse
  centralizing MR response parsing, EncodeRequest bounds check.
LogixDispatcher: symbol cache for O(1) tag lookup on repeated reads.
TagServices: ArrayPool for read response buffers and RMW temp buffer.
MultiServiceHandler: shared ArrayPool encode buffer across sub-responses.
EipScanner: use MrCodec.TryParseResponse, EncapsulationStatus enum.
Tests: updated for new error codes and type-safe Status comparisons.

+++ src/EipSim.Protocol/EipAdapter.cs
@@ -293,7 +293,7 @@ public sealed class EipAdapter : IAsyncDisposable
         {
             Command = req.Command,
             SessionHandle = req.SessionHandle,
-            Status = (uint)status,
+            Status = status,
             SenderContext = req.SenderContext,
         };
         var buf = new byte[EncapsulationHeader.Size];
